==> 0. esg_score_collection.ipynb <==
import yfinance as yf
from pprint import pprint
import pandas as pd
import pandas as pd
from tqdm import tqdm
from IPython.display import clear_output
import bs4 as bs
import requests
import yfinance as yf
import datetime

resp = requests.get('http://en.wikipedia.org/wiki/List_of_S%26P_500_companies')
soup = bs.BeautifulSoup(resp.text, 'lxml')
table = soup.find('table', {'class': 'wikitable sortable'})
tickers = []
for row in table.findAll('tr')[1:]:
    ticker = row.findAll('td')[0].text
    tickers.append(ticker)

tickers = [s.replace('\n', '') for s in tickers]
#start = datetime.datetime(2019,1,1)
#end = datetime.datetime(2019,7,17)
#data = yf.download(tickers, start=start, end=end)
#print(data)

f = pd.DataFrame(index = tickers, columns = ["socialScore","governanceScore", "environmentScore", "totalEsg", 'percentile','marketCap','dividendRate','fiftyDayAverage', 'shortRatio', 'revenueGrowth', 'debtToEquity','sector', 'industry'])

for i, ticker in tqdm(enumerate(tickers)):
    temp = yf.Ticker(ticker)
    try:
        social = temp.sustainability.loc['socialScore'][0]
        gov = temp.sustainability.loc['governanceScore'][0]
        env = temp.sustainability.loc['environmentScore'][0]
        total = temp.sustainability.loc['totalEsg'][0]
        perc = temp.sustainability.loc['percentile'][0]
        tl1 = [social, gov, env, total, perc]
        dict1 = temp.get_info()
        terms = ['marketCap','dividendRate','fiftyDayAverage', 'shortRatio', 'revenueGrowth', 'debtToEquity','sector', 'industry']
    
        tl = []
        for r in terms:
            tl.append(dict1[r])
              
        tl2 = tl1+tl
    except:
        continue
    f.iloc[i] = tl2
    

f.to_excel('data/ESG.xlsx')
