==> 4. regression_within_sector.ipynb <==
import numpy as np
import pandas as pd

from time import gmtime, strftime
import time
import os
from tqdm import tqdm
import pickle
from scipy.stats import linregress
from collections import defaultdict
from sklearn.feature_extraction.text import TfidfTransformer, CountVectorizer

from utils.crawler import Scrape10K, Scrape10Q, ConvertHTML
from utils.preprocessing import get_texts, stop_words

def get_cik(ticker):
    """ Get the cik for the ticker specified by the input argument 
    Input:
        ticker(str): ticker of the company e.g. "FB"
    """
    return ticker_library[ticker_library.ticker == ticker].secfilings.values[0][-10:]

def get_ciks(tickers):
    ciks = []

    for ticker in tickers:
        ciks.append(get_cik(ticker))

    return ciks

sector = "Energy"
# sector = "Energy"

# score_type = "governanceScore"
score_type = "environmentScore"

df_esg_score = pd.read_excel("data/esg_score.xlsx", sheet_name = "data")
tickers = df_esg_score[df_esg_score["sector"] == sector]["Company"]
esgs = df_esg_score[df_esg_score["sector"] == sector][["Company", "socialScore", "governanceScore", "environmentScore"]]

esgs.head()

print("Number of companies in Energy sector: {}".format(len(tickers)))

ticker_library = pd.read_csv(os.path.join("data", "tickers.csv"))

ciks = []
for ticker in tickers:    
    try:
        # for a given ticker, find its cik number through th ticker library
        ciks.append(ticker_library[ticker_library.ticker == ticker].secfilings.values[0][-10:])
    except:
        # if could not find cik, give it a empty cik
        ciks.append('')

# Analysis based on Energy sector

## Scrape Energy sector

browse_url_base_10k = 'https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=%s&type=10-K'
filing_url_base = 'http://www.sec.gov/Archives/edgar/data/%s/%s-index.html'
doc_url_base = 'http://www.sec.gov/Archives/edgar/data/%s/%s/%s'

browse_url_base_10q = 'https://www.sec.gov/cgi-bin/browse-edgar?action=getcompany&CIK=%s&type=10-Q&count=1000'

# Initialize log file
# (log file name = the time we initiate scraping session)
t = strftime("%Y_%m_%d_%H_%M_%S", gmtime())
log_file_name = t + ".txt"
log_file_path = os.path.join("log", log_file_name)

with open(log_file_path, 'a') as log_file:
    log_file.close()

# Iterate over CIKs and scrape 10-Ks
for cik in tqdm(ciks):
    time.sleep(5)
    Scrape10K(browse_url_base=browse_url_base_10k, 
          filing_url_base=filing_url_base, 
          doc_url_base=doc_url_base, 
          cik=cik,
          log_file_name=log_file_path)

    time.sleep(5)
    Scrape10Q(browse_url_base = browse_url_base_10q, 
          filing_url_base = filing_url_base, 
          doc_url_base = doc_url_base, 
          cik = cik,
          log_file_name = log_file_path)

#return to the main menu

## Transform from HTML to txt

# For 10-Ks...
# -*- coding: utf-8 -*-
dir_10k = os.path.join("data", "10k")
dir_10q = os.path.join("data", "10q")

# Iterate over CIKs and clean HTML filings
for cik in tqdm(ciks):
    ConvertHTML(cik, dir_10k)

# For 10-Qs...
# Iterate over CIKs and clean HTML filings
for cik in tqdm(ciks):
    ConvertHTML(cik, dir_10q)

## Get the preprocessed texts

ret = get_texts(ciks, tickers)
docs = ret["docs"]

# Regression of E, S, G score on indicator function(occurence of word_i)

Company - word table: 1 if word occurs in a certain doc, else 0

counters = defaultdict(list)

for ticker in tickers.values:
    for file_typ in ["10k", "10q"]:
        counter_path = os.path.join("data", file_typ, get_cik(ticker), "pickle", "token_counter.pkl")
        with open(counter_path, 'rb') as handle:
            counter = pickle.load(handle)
            counters[ticker].extend(list(counter.keys()))

dictionary = list()
for ticker in tickers.values:
    dictionary.extend(counters[ticker])

dictionary = set(dictionary)
print("Number of distinct words: {}".format(len(dictionary)))

dictionary.remove('')

word_count = defaultdict(int)
for word in dictionary:
    cnt = 0
    for ticker in tickers:
        cnt += word_count.get(word, 0)
    word_count[word] += cnt
    
top_k_word = [k for k, v in sorted(word_count.items(), key=lambda item: item[1], reverse=True)][:5000]

df_ind = pd.DataFrame(columns=["social_beta", "governance_beta", "environment_beta"], index = top_k_word)


for typ in ["social", "governance", "environment"]:
    score = esgs["{}Score".format(typ)]
    slopes = []
    for word in top_k_word:
        inds = []
        for ticker in tickers:
            if word in counters[ticker]:
                inds.append(1)
            else:
                inds.append(0)
        slope, intercept, *_ = linregress(inds, score)
        slopes.append(slope)
    df_ind["{}_beta".format(typ)] = slopes

df_ind = df_ind.sort_values(by=["social_beta", "governance_beta", "environment_beta"], ascending=False)

df_ind.head()

df_ind.to_csv(os.path.join("data", "reg_result_ind.csv"))

with open("dict_lemmatize.txt", 'w') as f:
    f.write('\n'.join(list(energy_dictionary)))

# Regression of E, S, G score on tfidf score for each word in each doc

cv = CountVectorizer(max_df=0.8, stop_words=stop_words, max_features=1000)
word_count_vector = cv.fit_transform(docs)

tfidf_transformer = TfidfTransformer(smooth_idf=True, use_idf=True)
tfidf_transformer.fit(word_count_vector)

# top 10000 frequent words
feature_names = cv.get_feature_names()

print("Number of words: {}".format(len(feature_names)))

df_doc_word = pd.DataFrame(columns=feature_names, index=tickers)

for i, ticker in tqdm(enumerate(tickers)):
    tf_idf_vector = tfidf_transformer.transform(cv.transform([docs[i]]))
    
    coo_matrix = tf_idf_vector.tocoo()
    # coo_matrix: A sparse matrix in which coo_matrix.col stores word_idx, coo_matrix.data stores tfidf score
    
    tuples = zip(coo_matrix.col, coo_matrix.data)
    for word_idx, tfidf in tuples:
        df_doc_word.at[ticker, feature_names[word_idx]] = tfidf

df_doc_word = df_doc_word.fillna(0)

df_tfidf = pd.DataFrame(columns=["social_beta", "governance_beta", "environment_beta"], index=feature_names)

for typ in ["social", "governance", "environment"]:
    score = esgs["{}Score".format(typ)]
    slopes = []
    for word in feature_names:
        tfidfs = df_doc_word[word].values.astype(float)
        slope, intercept, *_ = linregress(tfidfs, score)
        slopes.append(slope)
    df_tfidf["{}_beta".format(typ)] = slopes

cols = df_tfidf.columns
alpha = 0.3

for col in cols: 
    betas = df_tfidf[col]
    score_type = col.split('_')[0]
    
    upper_score = np.quantile(betas, 1 - alpha)
    lower_score = np.quantile(betas, alpha)
    is_good = np.where(betas < lower_score, 1, 0) + np.where(betas > upper_score, -1, 0)
    
    df_tfidf["{}_isGood".format(score_type)] = is_good

df_tfidf

Use just the top 5000 words, same as the 5000 words in indicator mode

# word_to_drop = list(set(feature_names) - set(top_k_word))

df_tfidf = df_tfidf.sort_values(by=["social_beta", "governance_beta", "environment_beta"], ascending=False)

df_tfidf = pd.read_csv("data/reg_result_tfidf.csv", index_col=0)

# "governance_beta", "environment_beta"
df_tfidf.sort_values(by=["social_beta"], ascending=False)

## Model Wrapup

not_listed = ["BBWI"]
sectors = ['Consumer Cyclical', 'Energy', 'Industrials', 'Healthcare',
       'Basic Materials', 'Consumer Defensive', 'Utilities', 'Technology',
       'Financial Services', 'Communication Services', 'Real Estate']

sector = "Financial Services"
# sector = "Energy"

df_esg_score = pd.read_excel("data/esg_score.xlsx", sheet_name = "data")
esgs = df_esg_score[df_esg_score["sector"] == sector][["Company", "socialScore", "governanceScore", "environmentScore"]]

tickers = list(esgs["Company"])
for t in not_listed:
    if t in tickers:
        tickers.remove(t)
        esgs = esgs.drop(esgs[esgs["Company"] == t].index)

ticker_library = pd.read_csv(os.path.join("data", "tickers.csv"))

ciks = get_ciks(tickers)

ret = get_texts(ciks, tickers)
docs = ret["docs"]

cv = CountVectorizer(max_df=0.8, stop_words=stop_words, max_features=1000)
word_count_vector = cv.fit_transform(docs)

tfidf_transformer = TfidfTransformer(smooth_idf=True, use_idf=True)
tfidf_transformer.fit(word_count_vector)

feature_names = cv.get_feature_names()

df_doc_word = pd.DataFrame(columns=feature_names, index=tickers)

for i, ticker in tqdm(enumerate(tickers)):
    tf_idf_vector = tfidf_transformer.transform(cv.transform([docs[i]]))
    
    coo_matrix = tf_idf_vector.tocoo()
    # coo_matrix: A sparse matrix in which coo_matrix.col stores word_idx, coo_matrix.data stores tfidf score
    
    tuples = zip(coo_matrix.col, coo_matrix.data)
    for word_idx, tfidf in tuples:
        df_doc_word.at[ticker, feature_names[word_idx]] = tfidf

df_doc_word = df_doc_word.fillna(0)

df_tfidf = pd.DataFrame(columns=["social_beta", "governance_beta", "environment_beta"], index=feature_names)

for typ in ["social", "governance", "environment"]:
    score = esgs["{}Score".format(typ)]
    slopes = []
    for word in feature_names:
        tfidfs = df_doc_word[word].values.astype(float)
        slope, intercept, *_ = linregress(tfidfs, score)
        slopes.append(slope)
    df_tfidf["{}_beta".format(typ)] = slopes

cols = df_tfidf.columns
alpha = 0.3

for col in cols: 
    betas = df_tfidf[col]
    score_type = col.split('_')[0]
    
    upper_score = np.quantile(betas, 1 - alpha)
    lower_score = np.quantile(betas, alpha)
    is_good = np.where(betas < lower_score, 1, 0) + np.where(betas > upper_score, -1, 0)
    
    df_tfidf["{}_isGood".format(score_type)] = is_good

df_tfidf.to_csv("data/tfidf_scores/{}.csv".format(sector))

## Appendix

import shutil

for cik in ciks:
    pkl_path = os.path.join("data", "10k", cik, "pickle")
    shutil.rmtree(pkl_path)
    pkl_path = os.path.join("data", "10q", cik, "pickle")
    shutil.rmtree(pkl_path)
