==> 5. good_vs_bad.ipynb <==
import pandas as pd
import numpy as np
import os
from sklearn.feature_extraction.text import TfidfTransformer, CountVectorizer
from tqdm import tqdm
from sklearn.metrics import confusion_matrix

from utils.preprocessing import get_texts, stop_words

# TODO: Modify here for different sectors and score_types
# set the sector and ESG type for the analysis
sector = "Consumer Cyclical"    # ['Consumer Cyclical', 'Energy', 'Industrials', 'Healthcare', 'Basic Materials', 'Consumer Defensive', 'Utilities', 'Technology', 'Financial Services', 'Communication Services', 'Real Estate']

score_type = "socialScore"  # ['socialScore', 'governanceScore', 'environmentScore']

df_esg_score = pd.read_excel("data/esg_score.xlsx", sheet_name = "data")
df_esg_score["sector"].dropna().unique()

## Get texts for companies

tickers = df_esg_score[df_esg_score["sector"] == sector]["Company"]
esgs = df_esg_score[df_esg_score["sector"] == sector][["Company", "socialScore", "governanceScore", "environmentScore"]]

len(tickers)

score = esgs[score_type]

alpha = 0.3
upper_score = np.quantile(score, 1 - alpha)
lower_score = np.quantile(score, alpha)

bad_companies = esgs[esgs[score_type] > upper_score]["Company"].values
good_companies = esgs[esgs[score_type] < lower_score]["Company"].values

good_companies

bad_companies

esgs[esgs[score_type] > upper_score].head()

bad_companies_score = esgs[esgs[score_type] > upper_score][score_type].values
good_companies_score = esgs[esgs[score_type] < lower_score][score_type].values

avg_bad = np.mean(bad_companies_score)
avg_good = np.mean(good_companies_score)
print(avg_bad, avg_good)

print(upper_score, lower_score)

ticker_library = pd.read_csv(os.path.join("data", "tickers.csv"))
good_cik = []
bad_cik = []
for ticker in good_companies:
    try:
        # for a given ticker, find its cik number through th ticker library
        good_cik.append(ticker_library[ticker_library.ticker == ticker].secfilings.values[0][-10:])
    except:
        # if could not find cik, give it a empty cik
        good_cik.append('')

for ticker in bad_companies:    
    try:
        # for a given ticker, find its cik number through th ticker library
        bad_cik.append(ticker_library[ticker_library.ticker == ticker].secfilings.values[0][-10:])
    except:
        # if could not find cik, give it a empty cik
        bad_cik.append('')

ret_good = get_texts(good_cik, good_companies)

ret_bad = get_texts(bad_cik, bad_companies)

good_docs = ret_good["docs"]
bad_docs = ret_bad["docs"]

# TODO: Modify here for different ngram range
n_min = 2
n_max = 3
cv = CountVectorizer(max_df=0.8, stop_words=stop_words, max_features=200, ngram_range=(n_min, n_max))
word_count_vector = cv.fit_transform(good_docs + bad_docs)
feature_names = cv.get_feature_names()

count_feature = word_count_vector.toarray().sum(axis=0)

feature_names = cv.get_feature_names()

d = {"word": [], "good_score": [], "bad_score": [], "good_score_all": []
    , "bad_score_all": [], "count": [], "good_nums": [], "bad_nums": []}

for feature_idx, word in enumerate(feature_names):
    good_sum = bad_sum = good_num = bad_num = 0

    for i, doc_set in enumerate(good_docs):
        if word in doc_set:
            good_num += 1
            good_sum += good_companies_score[i]
    for i, doc_set in enumerate(bad_docs):
        if word in doc_set:
            bad_num += 1
            bad_sum += bad_companies_score[i]
    
    # print("word: {}".format(word))
    d["word"].append(word) 
    
    if good_num:
        d["good_score"].append(good_sum / good_num)
    else:
        d["good_score"].append(0)
    if bad_num:
        d["bad_score"].append(bad_sum / bad_num)
    else:
        d["bad_score"].append(0)

    d["good_score_all"].append(good_sum / len(good_docs))
    d["bad_score_all"].append(bad_sum / len(bad_docs))

    d["count"].append(count_feature[feature_idx])
    d["good_nums"].append(good_num)
    d["bad_nums"].append(bad_num)


df = pd.DataFrame(data=d)

df.head()

df["diff"] = abs(df["good_nums"] - df["bad_nums"])

df = df.sort_values("diff", ascending=False)#.head(60)

df.head(10)

goodvbad_path = os.path.join("data", "goodvbad")
if not os.path.isdir(goodvbad_path):
    os.mkdir(goodvbad_path)

df.round(2).to_csv("data/goodvbad/{}_{}_{}_n{}-{}.csv".format(sector[:8], score_type[:3], alpha, n_min, n_max))

## Appendix

# To remove all the previous files earlier than 2020
for cik in tqdm(bad_cik):
    tenk_path = os.path.join("data", "10k", cik, "rawtext")
    all_raws = os.listdir(tenk_path)
    for filename in all_raws:
        if filename[0] == '.':
            continue
        year = int(filename.split('_')[1].split('-')[0])
        # print(year)
        if year < 2020:
            # print(os.path.join(tenk_path, filename))
            os.remove(os.path.join(tenk_path, filename))
    
    tenq_path = os.path.join("data", "10q", cik, "rawtext")
    all_raws = os.listdir(tenq_path)
    for filename in all_raws:
        if filename[0] == '.':
            continue
        year = int(filename.split('_')[1].split('-')[0])
        # print(year)
        if year < 2020:
            # print(os.path.join(tenk_path, filename))
            os.remove(os.path.join(tenq_path, filename))


# Remove all the previous incompleted scraping
dirname = os.path.join("data", "10q")
all_files = os.listdir(dirname)

import shutil
for filename in all_files:
    pkldir = os.path.join(dirname, filename, "pickle")
    if os.path.isdir(pkldir):
        if not os.path.exists(os.path.join(pkldir, "agg_texts.pkl")):
            shutil.rmtree(os.path.join(dirname, filename))

df = pd.read_csv("energy_good_vs_bad_uni_bi_tri.csv", index_col=0)

df.head()

comp = pd.read_csv("data/sp500_component_stocks.csv")


comp[comp['A'] == "C"]["Agilent Technologies Inc."]

comp[comp['A'].isin(good_companies)]["Agilent Technologies Inc."]

comp[comp['A'].isin(bad_companies)]["Agilent Technologies Inc."]

ticker_library = pd.read_csv(os.path.join("data", "tickers.csv"))

## Helper functions

def get_cik(ticker):
    """ Get the cik for the ticker specified by the input argument 
    Input:
        ticker(str): ticker of the company e.g. "FB"
    """
    return ticker_library[ticker_library.ticker == ticker].secfilings.values[0][-10:]
    

def get_ciks(tickers):
    ciks = []

    for ticker in tickers:
        ciks.append(get_cik(ticker))

    return ciks

def ngrams(s, n):
    """ Get all the n-gram for input texts s
    Input:
        s (str): A string of texts with each word separated by a whitespace
        n (int): n-gram to extract
    Return:
        [str]: A list of string in the following format ([['a', 'b'], ['b', 'c'], ['c', 'd']])
    """
    
    s = s.split(' ')
    output = []
    for i in range(len(s) - n + 1):
        output.append(s[i:i+n])

    return output

# ngrams('a b c d', 2) # [['a', 'b'], ['b', 'c'], ['c', 'd']]

## Validation codes

def get_count(doc, df_dict, n_min, n_max):
    """ Count the number of good and bad words occurred in the document
    Input:
        doc (str): A string with all the words in the documents
        df_dict (pd.DataFrame): A DataFrame with word and isGood column, generated by previous section
        n_min, n_max (int): specify the ngram range used to generate the dictionary, should be consistent with how df_dict is generated
    Return:
        (dict): A dictionary with value good_count and bad_count
    """
    grams = []
    for n in range(n_min, n_max + 1):
        grams.extend([' '.join(li) for li in ngrams(doc, 2)])
    
    good_count = bad_count = 0
    
    for g in grams:
        if g in df_dict["word"].values:
            val = df_dict[df_dict["word"] == g]["isGood"].values
            if val == 1:
                good_count += 1
            elif val == -1:
                bad_count += 1

    return {"good_count": good_count, "bad_count": bad_count}
    

def validation(df_topk, val_tickers, dict_threshold):
    """ Perform the validation step
    The validation rationale: Companies whose score are in upper 50% group are considered "bad" companies and the corresponding val_true = 1; 0 otherwise (in lower 50% group, which is considered a good company)
    Input:
        df_topk (pd.DataFrame): containes the sector specific dict
        val_tickers (list): A list of tickers to be validated
        dict_threshold (int): A gram is considered as a good word if its good_nums - bad_nums > threshold
        val_true (list): A list of true labels for each companies
    """
    diff = df_topk["good_nums"] - df_topk["bad_nums"]
    df_topk["isGood"] = diff.apply(lambda x: 1 if x > dict_threshold else (
        -1 if x < -dict_threshold else 0))

    # 1 if good_nums - bad_nums > threshold; -1 if good_nums - bad_nums < -threshold; 0 otherwise
    val_ciks = [get_cik(ticker) for ticker in val_tickers]
    
    ret_texts = get_texts(val_ciks, val_tickers) 

    val_pred = []
    for doc in tqdm(ret_texts["docs"]):
        ret = get_count(doc, df_topk[["word", "isGood"]], 2, 3)

        if ret["good_count"] - ret["bad_count"] > 0:
            val_pred.append(1)
        else:
            val_pred.append(0)
    
    print("val_pred: {}".format(val_pred))
    
    return val_pred

bad_companies

# tickers = ['NOV', 'OKE', 'HAL', 'SLB', 'WMB', 'KMI']
tickers = list(good_companies) + list(bad_companies)
ciks = [get_cik(ticker) for ticker in tickers]

ret = get_texts(ciks, tickers) 

val_true = [1 for _ in good_companies] + [0 for _ in bad_companies]

(df["good_nums"] > df["bad_nums"]).sum()

len(df)

val_pred = validation(df, tickers, 1)

cm = confusion_matrix(val_true, val_pred)
print("Confusion Matrix: \n{}".format(cm))

ddddf = pd.read_csv("data/goodvbad/env_0.333_n2-3.csv", index_col=0)

ddddf.head()

num_good = (ddddf["good_nums"] > ddddf["bad_nums"]).sum()
num_bad = (ddddf["good_nums"] < ddddf["bad_nums"]).sum()
print(num_good, num_bad)

total_num_good = ddddf["good_nums"].sum()
total_num_bad = ddddf["bad_nums"].sum()
print(total_num_good, total_num_bad)

dict_threshold = 1
diff = df["good_nums"] - df["bad_nums"]
df["isGood"] = diff.apply(lambda x: 1 if x > dict_threshold else (
    -1 if x < -dict_threshold else 0))

df.head()

df[["word", "isGood"]]

## Generate N-grams based on sectors

ticker_library = pd.read_csv(os.path.join("data", "tickers.csv"))

sector = "Energy"
# sector = "Energy"

# score_type = "governanceScore"
score_type = "environmentScore"

sectors = ['Consumer Cyclical', 'Technology', 'Financial Services']
score_types = ["governanceScore", "environmentScore", "socialScore"]

not_listed = ["BBWI"]

with open('data/n_grams/Technology_environmentScore.pkl', 'rb') as handle:
    feature_names = pickle.load(handle)

feature_names

import pickle

df_esg_score = pd.read_excel("data/esg_score.xlsx", sheet_name = "data")

for sector in sectors:
    for score_type in score_types:
        esgs = df_esg_score[df_esg_score["sector"] == sector][["Company", "socialScore", "governanceScore", "environmentScore"]]
        score = esgs[score_type]
        alpha = 0.3
        upper_score = np.quantile(score, 1 - alpha)
        lower_score = np.quantile(score, alpha)

        bad_companies = list(esgs[esgs[score_type] > upper_score]["Company"].values)
        good_companies = list(esgs[esgs[score_type] < lower_score]["Company"].values)

        for t in not_listed:
            if t in bad_companies:
                bad_companies.remove(t)
            if t in good_companies:
                good_companies.remove(t)

        tickers = good_companies + bad_companies
        ciks = get_ciks(tickers)
        ret = get_texts(ciks, tickers)
        docs = ret["docs"]
        
        n_min = 2
        n_max = 3
        cv = CountVectorizer(max_df=0.7, stop_words=stop_words, max_features=200, ngram_range=(n_min, n_max))
        word_count_vector = cv.fit_transform(docs)
        feature_names = cv.get_feature_names()

        with open('data/n_grams/{}_{}.pkl'.format(sector, score_type), 'wb') as handle:
            pickle.dump(feature_names, handle, protocol=pickle.HIGHEST_PROTOCOL)


with open('data/n_grams/{}_{}.pkl'.format(sector, score_type), 'wb') as handle:
    pickle.dump(feature_names, handle, protocol=pickle.HIGHEST_PROTOCOL)

tickers = df_esg_score[df_esg_score["sector"] == sector]["Company"]
esgs = df_esg_score[df_esg_score["sector"] == sector][["Company", "socialScore", "governanceScore", "environmentScore"]]
ciks = get_ciks(tickers)

ret = get_texts(ciks, tickers)
docs = ret["docs"]

# TODO: Modify here for different ngram range
n_min = 2
n_max = 3
cv = CountVectorizer(max_df=0.6, stop_words=stop_words, max_features=200, ngram_range=(n_min, n_max))
word_count_vector = cv.fit_transform(docs)
feature_names = cv.get_feature_names()

d = {"combination": [], "bad_tickers": [], "good_tickers": []}

for sec in df_esg_score["sector"].dropna().unique():
    for s_type in ["governanceScore", "environmentScore", "socialScore"]:
        tickers = df_esg_score[df_esg_score["sector"] == sec]["Company"]
        esgs = df_esg_score[df_esg_score["sector"] == sec][["Company", "socialScore", "governanceScore", "environmentScore"]]
        score = esgs[s_type]
        alpha = 0.25
        upper_score = np.quantile(score, 1 - alpha)
        lower_score = np.quantile(score, alpha)

        bad_companies = esgs[esgs[s_type] > upper_score]["Company"].values
        good_companies = esgs[esgs[s_type] < lower_score]["Company"].values

        # print(sec)
        d["combination"].append(sec + "_" + s_type)
        d["good_tickers"].append(','.join(good_companies))
        d["bad_tickers"].append(','.join(bad_companies))

ddf = pd.DataFrame(data=d)
ddf.to_csv("data/goodvbad/goodcbad_companies.csv")
